Replace Sail for the remaining functions in the vector crypto Sail supplement.
The `rol32` in the supplement gets mapped to an overload of `<<<` implemented
by `rotate_bits_left`.  It seems pointless to add `rotate_bits_left` to the
supplement since it doesn't show up directly in the Sail source for the
affected instructions.

With the hyperlinked source, the supplement is superfluous anyway.