Add Sail wavedrom diagrams for some instruction encodings.
The original diagrams have been left in place for easier
comparison.

The main difference between the two occurs- when the manual
has separate diagrams for RV32 and RV64, and Sail has a single
clause with arch differences handled in `when` clause.

There are occasional infelicities when a Sail variable does
not fit into the bitwidth of the diagram.